some bug fixed (non-gpu job kill) add priotize after submit job add max job per user limit (default 12)

+++ gorque.py
@@ -6,6 +6,18 @@ import sqlite3
 import time
 import sys
 import getopt
+import json
+
+
+def load_config():
+    # for future usage
+    f = open('config.json')
+    ls = f.readlines()
+    j = ''
+    for l in ls:
+        if not l.strip().startswith('//'):
+            j = j + l
+    return json.loads(j)
 
 
 def humanize_time(secs):
@@ -36,6 +48,7 @@ class queue:
                     mode TEXT, node TEXT, name TEXT, pid INTEGER)''')
             conn.commit()
         self.conn = conn
+        self.max_job_per_user_limit = 12
 
     def print_queue(self, all=False):
         # print the current queue
@@ -67,7 +80,6 @@ class queue:
             print template.format(str(row[7]), crop_string(row[0], 20), row[1], str(row[2]), time_passed, mode, row[6])
         print template.format('-' * 4, '-' * 20, '-' * 10, '-' * 10, '-' * 10, '-' * 6, '-' * 13)
         print '| ' + str(job_count) + ' jobs running now.'
-
 
     def insert_job(self, name, user, priority, script):
         c = self.conn.cursor()
@@ -132,6 +144,11 @@ class queue:
         c.execute('''UPDATE queue SET mode = 'F', end_time = ? WHERE ROWID = ?''', (int(time.time()), str(rowid)))
         self.conn.commit()
 
+    def prioritize_job(self, rowid, priority):
+        c = self.conn.cursor()
+        c.execute('''UPDATE queue SET priority = ? WHERE ROWID = ?''', (int(priority), str(rowid)))
+        self.conn.commit()
+
     def kill_job(self, rowid):
         c = self.conn.cursor()
         c.execute('''SELECT pid, mode, node FROM queue WHERE ROWID = ?''', (str(rowid),))
@@ -141,8 +158,13 @@ class queue:
             os.system('kill -9 ' + str(job[0]))
             os.system('kill -9 ' + str(job[0] + 1))
             gpu_pid = subprocess.check_output('''ssh %s 'nvidia-smi | sed "16!d" | tr -s " " | cut -d" " -f3 ' ''' % (job[2],), shell=True)
-            print gpu_pid
-            os.system('''ssh %s 'kill -9 %s' ''' % (job[2], gpu_pid))
+            if gpu_pid.strip() == 'running':
+                # no GPU job
+                pass
+            else:
+                # kill GPU job
+                print gpu_pid
+                os.system('''ssh %s 'kill -9 %s' ''' % (job[2], gpu_pid))
         c = self.conn.cursor()
         c.execute('''UPDATE queue SET mode = 'K', end_time = ?, pid = NULL WHERE ROWID = ?''', (int(time.time()), str(rowid)))
         self.conn.commit()
@@ -163,11 +185,22 @@ class queue:
         free_nodes = [x for x in free_nodes.split('\n') if x != '']
         return free_nodes
 
-    def check_jobs(self):
+    def get_waiting_jobs(self):
         c = self.conn.cursor()
-        c.execute('''SELECT ROWID FROM queue WHERE mode = 'Q' ORDER BY priority DESC ''')
+        c.execute('''SELECT ROWID, user FROM queue WHERE mode = 'Q' ORDER BY priority DESC ''')
         waiting_jobs = c.fetchall()
-        if len(waiting_jobs) > 0:
+        qualified_jobs = []
+        for job in waiting_jobs:
+            c.execute('''SELECT ROWID FROM queue WHERE mode = 'R' AND user = ?''', (job[1],))
+            user_jobs = c.fetchall()
+            if len(user_jobs) < self.max_job_per_user_limit:
+                qualified_jobs.append(job)
+        return qualified_jobs
+
+    def check_jobs(self):
+        qualified_jobs = self.get_waiting_jobs()
+        if len(qualified_jobs) > 0:
+            # make sure there is waiting jobs then query
             free_nodes = self.find_free_nodes()
             if len(free_nodes) > 0:
                 # last use compute-0-0
@@ -175,13 +208,16 @@ class queue:
                     # compute0 = free_nodes[0]
                     free_nodes.remove(free_nodes[0])
                     # free_nodes.append(compute0)
-                # refetch database in case already executed by another cron process.
-                c.execute('''SELECT ROWID FROM queue WHERE mode = 'Q' ORDER BY priority DESC ''')
-                waiting_jobs = c.fetchall()
-                if len(waiting_jobs) > 0:
-                    job = waiting_jobs[0]
+                # refetch database in case already
+                # executed by another cron process.
+                qualified_jobs = self.get_waiting_jobs()
+                if len(qualified_jobs) > 0:
+                    # check user limit
+                    job = qualified_jobs[0]
                     node = free_nodes[0]
                     self.execute_job(job[0], node)
+        else:
+            print 'no qualified waiting jobs in the queue'
 
 
 def main(argv):
@@ -189,7 +225,7 @@ def main(argv):
     script_path = None
     delete_rowid = None
     try:
-        opts, args = getopt.getopt(argv, "u:s:lad:c")
+        opts, args = getopt.getopt(argv, "u:s:lad:cp:")
     except getopt.GetoptError:
         print 'wrong syntax, please contact admin to request how to manual.'
         sys.exit(2)
@@ -198,6 +234,11 @@ def main(argv):
             user = arg
         elif opt in ("-s"):
             script_path = arg
+        elif opt in ("-p"):
+            job_id = arg
+            priority = raw_input('input the priority: ')
+            q = queue()
+            q.prioritize_job(job_id, priority)
         elif opt in ("-l"):
             q = queue()
             q.print_queue()
